% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-backend.R
\name{add_queries_to_group}
\alias{add_queries_to_group}
\title{add_queries_to_group}
\usage{
add_queries_to_group(group)
}
\arguments{
\item{group}{a group as returned by the set_to_gorup function}
}
\value{
a tibble with binary expansions of terms and English queries of them
}
\description{
add_queries_to_group
}
