% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{create_eric_url}
\alias{create_eric_url}
\title{Database Communications}
\usage{
create_eric_url(
  search_terms,
  start = 0,
  rows = 0,
  fields = list(list("*"), list("peerreviewed", "'T'"))
)
}
\arguments{
\item{search_terms}{.}

\item{start}{.}

\item{rows}{.}

\item{fields}{.}
}
\value{
string of html encoded eric query
}
\description{
This file contains functions to communicate with the ERIC and Proquest databases
}
