% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{create_heatmap}
\alias{create_heatmap}
\title{create_heatmap}
\usage{
create_heatmap(explor, sa, sb)
}
\arguments{
\item{explor}{the explor tibble to construct the heatmap out of}

\item{sa}{the first set of terms to compare}

\item{sb}{the second set of terms to compare}
}
\value{
a ggplot plot showing a heatmap correlation between the sets
}
\description{
create_heatmap
}
