% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{example_explor_b}
\alias{example_explor_b}
\title{An example explor tibble}
\format{
A tibble with four search terms, queries, eric counts, and proquest counts. NA represents data has not been collected
\describe{
   \item{A}{search term "A"}
   \item{B}{search term "B"}
   \item{C}{search term "C"}
   \item{D}{search term "D"}
   \item{query}{English translation of search query}
   \item{eric}{Number of matching articles in ERIC}
   \item{proquest}{Number of matching articles in Proquest}
}
}
\usage{
example_explor_b
}
\description{
A dataset containing a fully formed term search set,
English translations, and eric article counts
}
\keyword{datasets}
