% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_arxiv_count}
\alias{get_arxiv_count}
\title{get_arxiv_count}
\usage{
get_arxiv_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles arxiv has for the search
}
\description{
get_arxiv_count
}
