% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_asee_peer_count}
\alias{get_asee_peer_count}
\title{get_asee_peer_count}
\usage{
get_asee_peer_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles asee has for the search
}
\description{
get_asee_peer_count
}
