% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_count}
\alias{get_count}
\title{get_count}
\usage{
get_count(database, query)
}
\arguments{
\item{database}{a string indicating which database to query}

\item{query}{a query string, eg A OR B AND C}
}
\value{
an integer for the count of papers from a database
}
\description{
get_count
}
