% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_ebsco_count}
\alias{get_ebsco_count}
\title{get_ebsco_count}
\usage{
get_ebsco_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles ebsco has for the search
}
\description{
get_ebsco_count
}
