% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_eric_count}
\alias{get_eric_count}
\title{get_eric_count}
\usage{
get_eric_count(str)
}
\arguments{
\item{str}{}
}
\value{
an integer with the corresponding number of articles eric has for the search
}
\description{
get_eric_count
}
