% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_google_scholar_count}
\alias{get_google_scholar_count}
\title{get_google_scholar_count}
\usage{
get_google_scholar_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles google has for the search
}
\description{
get_google_scholar_count
}
