% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_jstor_count}
\alias{get_jstor_count}
\title{get_jstor_count}
\usage{
get_jstor_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles jstor has for the search
}
\description{
get_jstor_count
}
