% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_plos_count}
\alias{get_plos_count}
\title{get_plos_count}
\usage{
get_plos_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles plos has for the search
}
\description{
get_plos_count
}
