% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_proquest_count}
\alias{get_proquest_count}
\title{get_proquest_count}
\usage{
get_proquest_count(str)
}
\arguments{
\item{str}{terms to search for}
}
\value{
an integer with the corresponding number of articles proquest has for the search
}
\description{
get_proquest_count
}
