% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-backend.R
\name{group_to_explor}
\alias{group_to_explor}
\title{group_to_explor}
\usage{
group_to_explor(group)
}
\arguments{
\item{group}{a group as returned by the set_to_gorup function}
}
\value{
a tibble with binary expansions of terms and English queries of them
}
\description{
group_to_explor
}
