% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-backend.R
\name{make_set}
\alias{make_set}
\title{Data Backend}
\usage{
make_set(input_terms)
}
\arguments{
\item{input_terms}{a string in the form "A OR B OR C"}
}
\value{
A tibble with column names from \code{input_terms} representing search terms OR'd together
#make_set("A OR B")
#make_set("E OR F OR G")
#make_set("M OR N OR O OR P")
}
\description{
This file provides the model for combining search term sets into
a tibble for visualization.
}
