% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-backend.R
\name{set_to_group}
\alias{set_to_group}
\title{set_to_group}
\usage{
set_to_group(set, ...)
}
\arguments{
\item{set}{a list of term strings}

\item{...}{additional sets}

\item{i}{an integer}

\item{n}{the index of the bit to retreive}
}
\value{
a group tibble, representing search term sets which are AND'd together
set_to_group(make_set("A OR B"))
set_to_group(make_set("A OR B"), make_set("C OR D"))
generate_name_from_set

string with the terms deliminated by an underscore
nth_bit

the nth bit of an integer
}
\description{
set_to_group
}
