% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{translate_binary_to_label}
\alias{translate_binary_to_label}
\title{translate_binary_to_label}
\usage{
translate_binary_to_label(binary, sym)
}
\arguments{
\item{binary}{the binary columns from an explor tibble}

\item{sym}{array of term strings}
}
\value{
a string with the indicated symbols OR'd together
}
\description{
translate_binary_to_label
}
