% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-communication.R
\name{get_proquest_count}
\alias{get_proquest_count}
\title{get_proquest_count}
\usage{
get_proquest_count(
  terms,
  database = "medlineprof",
  authorization = "Bearer 8461684a-5a79-407c-9321-93d8d719ddeb"
)
}
\arguments{
\item{terms}{terms to search for}

\item{database}{proquest has lots of databases, which one are we after?}

\item{authorization}{our bearer token for access}
}
\value{
an integer with the corresponding number of articles proquest has for the search
}
\description{
get_proquest_count
}
