% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-backend.R
\name{group_to_explor}
\alias{group_to_explor}
\title{group_to_explor}
\usage{
group_to_explor(group_with_queries)
}
\arguments{
\item{group_with_queries}{a group as created by the add_queries_to_group function}
}
\value{
a fully expanded explor tibble without paper counts
group_to_explor(add_queries_to_group(set_to_group(make_set("A OR B"))))
group_to_explor(add_queries_to_group(set_to_group(make_set("A OR B"), make_set("C OR D"))))
}
\description{
group_to_explor
}
